Better average, gizmos, less movement
Added a players debugger to players so we can see the center of the
players' average positions as well as the two players who are furthest
away vertically and horizontally. Perlin shake is still up in the air.
Suggested using more subtle, less zoomy camera work as per advice from
"playtesters"