Moved the camera around, fixed the animator, fixing the animations, reset the origin
Oof. Moved everything to be more origin-centric. Added the old animator
back in, and started fixing the animations to be more cooperative with
the weird rotation that the model needs. Also fixed a crazy rotation in
the TennisSwing animation, and started working on fixing the
TennisStrike animation to fit in with the new rotation paradigm.