Info Screens and Controller Layouts in place
The different info screens are completed and the different controller
layouts have been added to each menu.  Things that are left to do:
- Labels for the cursors in ColorSelection
- Images for the buttons in LevelSelection (these will most likely be
placeholders because my version won’t have the completed games,
especially something like Tennis)
- Since Tennis is getting revamped, I left out its description for now

A major bug I’ve run into: going back from LevelSelection to
ColorSelection. This bug is happening because the colors that have been
picked aren’t being destroyed in the liaison when going backwards.
Something to think about…

The rest of the bugs are minor. I hopefully will figure out something
to fix when a player chooses the color Green and lands on the back
button in ColorSelection.  Right now, that player gets stuck.  He needs
to cycle through the nearest colors to see which one is available.

And…calling it a night because I’m exhausted.