Some odds and ends menu stuff
Fixed so we can now go back from Level Selection and reselect colors
without errors.  Also fixed a bug where colors confirmed would be
disabled, but still clickable.  Now we just have to figure out the
problem when someone is on green and another player is on the back
button